update connection DB
