+++ backend/src/app.ts
@@ -1,8 +1,7 @@
 import express from "express";
 import dotenv from "dotenv";
 import routes from "./routes";
-import dbConnect from "./db/dbConnect";
-import initializeDatabase  from "./init-db";
+import connectDB from "./db/dbConnect";
 
 dotenv.config();
 
@@ -18,11 +17,15 @@ const HOST = process.env.HOST || "localhost";
 // Rutas
 app.use("/api", routes);
 
-dbConnect().then(() => {
-    initializeDatabase(true);
-});
-
-// Iniciar el servidor
-app.listen(PORT, () => {
-    console.log(`Server is running at http://${HOST}:${PORT}`);
-});
+// Conexión a la base de datos y lanzamiento del servidor
+(async () => {
+    try {
+        await connectDB(); // Puedes pasar `true` si quieres forzar la inicialización
+        app.listen(PORT, () => {
+            console.log(`Server is running at http://${HOST}:${PORT}`);
+        });
+    } catch (error) {
+        console.error("Failed to start the server:", error);
+        process.exit(1);
+    }
+})();
