--- backend/src/init-db.ts
@@ -1,32 +1,0 @@
-import mongoose from "mongoose";
-import{ User }from "./models/user";
-import { Product } from "./models/products"
-
-async function initializeDatabase(force: boolean = false) {
-    try {
-        if (!mongoose.connection.readyState) {
-            throw new Error("Mongoose is not connected");
-        }
-
-        if (force) {
-            await mongoose.connection.dropDatabase().catch(error => {
-                throw new Error(`Error dropping database: ${error.message}`);
-            });
-            console.log("Database dropped.");
-        }
-
-        await User.init().catch(error => {
-            throw new Error(`Error initializing User collection: ${error.message}`);
-        });
-        
-        await Product.init().catch(error => {
-            throw new Error(`Error initializing Products collection: ${error.message}`);
-        });
-        
-        console.log("Database initialized, collections created.");
-    } catch (error) {
-        console.error("Error initializing database:", error);
-    }
-}
-
-export default initializeDatabase;
